+++ util/RiseFallMinMax.cc
@@ -34,8 +34,8 @@ RiseFallMinMax::RiseFallMinMax()
 void
 RiseFallMinMax::clear()
 {
-  for (int rf_index = 0; rf_index<RiseFall::index_count; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0; rf_index<RiseFall::index_count; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       exists_[rf_index][mm_index] = false;
     }
   }
@@ -43,8 +43,8 @@ RiseFallMinMax::clear()
 
 RiseFallMinMax::RiseFallMinMax(float init_value)
 {
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       values_[rf_index][mm_index] = init_value;
       exists_[rf_index][mm_index] = true;
     }
@@ -53,8 +53,8 @@ RiseFallMinMax::RiseFallMinMax(float init_value)
 
 RiseFallMinMax::RiseFallMinMax(const RiseFallMinMax *rfmm)
 {
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       values_[rf_index][mm_index] = rfmm->values_[rf_index][mm_index];
       exists_[rf_index][mm_index] = rfmm->exists_[rf_index][mm_index];
     }
@@ -154,8 +154,8 @@ RiseFallMinMax::setValue(const RiseFall *rf,
 void
 RiseFallMinMax::setValues(RiseFallMinMax *values)
 {
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       values_[rf_index][mm_index] = values->values_[rf_index][mm_index];
       exists_[rf_index][mm_index] = values->exists_[rf_index][mm_index];
     }
@@ -206,8 +206,8 @@ RiseFallMinMax::maxValue(// Return values
 {
   max_value = MinMax::max()->initValue();
   exists = false;
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       if (exists_[rf_index][mm_index]) {
         max_value = std::max(max_value, values_[rf_index][mm_index]);
         exists = true;
@@ -219,8 +219,8 @@ RiseFallMinMax::maxValue(// Return values
 bool
 RiseFallMinMax::empty() const
 {
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       if (exists_[rf_index][mm_index])
         return false;
     }
@@ -258,8 +258,8 @@ RiseFallMinMax::mergeWith(RiseFallMinMax *rfmm)
 bool
 RiseFallMinMax::equal(const RiseFallMinMax *values) const
 {
-  for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-    for (int mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
+  for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t mm_index = 0; mm_index < MinMax::index_count; mm_index++) {
       bool exists1 = exists_[rf_index][mm_index];
       bool exists2 = values->exists_[rf_index][mm_index];
       if (exists1 != exists2)
@@ -284,8 +284,8 @@ RiseFallMinMax::isOneValue(float &value) const
 {
   if (exists_[0][0]) {
     value = values_[0][0];
-    for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
-      for (int mm_index=0; mm_index<MinMax::index_count;mm_index++) {
+    for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+      for (size_t mm_index=0; mm_index<MinMax::index_count;mm_index++) {
         if (!exists_[rf_index][mm_index]
             || values_[rf_index][mm_index] != value)
           return false;
@@ -302,10 +302,10 @@ RiseFallMinMax::isOneValue(const MinMax *min_max,
                            // Return values.
                            float &value) const
 {
-  int mm_index = min_max->index();
+  size_t mm_index = min_max->index();
   if (exists_[0][mm_index]) {
     value = values_[0][mm_index];
-    for (int rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
+    for (size_t rf_index = 0 ; rf_index < RiseFall::index_count ; rf_index++) {
       if (!exists_[rf_index][mm_index]
           || values_[rf_index][mm_index] != value)
         return false;
@@ -316,4 +316,4 @@ RiseFallMinMax::isOneValue(const MinMax *min_max,
     return false;
 }
 
-} // namespace
+} // namespace sta
